taitogn.cpp : Move state class declaration into taitogn.cpp
